prepapring to submit data to backend
